==> 1 json_to_dataframe.ipynb <==
import pandas as pd
import glob
from multiprocessing import Pool
from multiprocessing import cpu_count
num_partitions = cpu_count()

# df = pd.read_json(r"/home/ubuntu/data/resolute/resolute/Achaia Aigeira_3313351.json", lines=True)
#scp -i ~/.ssh/ansible-user.pem revolut.ipynb ubuntu@10.91.52.118:/home/ubuntu/

The attributes: InternalPropertyDescription, ExternalPropertyDescription, ExtraInformation carry important information, however they are user based and most of them will likely to be misleading.

def json_to_dataframe(all_files_sample):
    column_list = ['Neighbourhood',
                   'Area_sqm', 'PropertyType',
                   'Bedrooms', 'Bathrooms', 'Rooms', 'Floor', 'Levels',
                   'PriceEuro','PricePerM2Euro',
                   'Latitude', 'Longitude',
                   'ConstructionYear', 'ModifiedOn']
    li = []
    
    for filename in all_files_sample:
        df = pd.read_json(filename, lines=True)[column_list]
        li.append(df)
    frame = pd.concat(li, axis=0, ignore_index=True)
    
    return frame

path = r"../data/resolute/resolute" # use your path
all_files = glob.glob(path + "/*.json")

all_files_split = np.array_split(all_files, num_partitions)

with Pool(processes=num_partitions) as pool1:
    super_frame = pd.concat(pool1.map(json_to_dataframe, all_files_split))

super_frame.head()

# convert to datetime object
super_frame['ModifiedOn'] = pd.to_datetime(super_frame['ModifiedOn'])

super_frame.dtypes

super_frame.sample()

### save to csv 

# super_frame.to_csv('df_resolute.csv', index=False)
