==> 2 third_party_geodata.ipynb <==
import libraries

import geopandas as gpd
import pandas as pd

Data

coastline: http://geodata.gov.gr/dataset

blue flag beaches: http://geodata.gov.gr/dataset/aktes-me-galazia-semaia-2010

cities: http://geodata.gov.gr/dataset/poleis

villages: http://geodata.gov.gr/dataset/oikismoi

census population: http://geodata.gov.gr/dataset/plethusmos-elladas

kap_dim = gpd.read_file('../data/oriaprokapodistriakwnota/oria_prokapodistriakwn_ota/oria_prokapodistriakwn_ota_WGS84_ALL.shp')\
#                 [['EKTASH','NAME_LATIN','HubDist','hubdist_2', 'hubdist_3','oiki_densi', 'medianpopu','geometry']]\
#                 .rename(columns={'HubDist':'dist_coast','hubdist_2':'dist_beach','hubdist_3':'dist_city', 'medianpopu':'pop_densi'})

# kap_dim.head()

enumerate the same NAME_LATIN

for i in kap_dim.NAME_LATIN.unique():
    if len(kap_dim[kap_dim.NAME_LATIN==i])>1:
        index_list = kap_dim[kap_dim.NAME_LATIN==i].index
        cnt=0
        for index in index_list:
            cnt+=1
            kap_dim.loc[index, 'NAME_LATIN'] ="{}_{}".format(kap_dim.loc[index, 'NAME_LATIN'], cnt)

# kap_dim[kap_dim.NAME_LATIN.str.contains('K. SKALAS VRVPOY')]

kap_dim.NAME_LATIN.nunique()

len(kap_dim)

# kap_dim.to_file(driver = 'ESRI Shapefile', filename= '../data/oriaprokapodistriakwnota/oria_prokapodistriakwn_ota/oria_prokapodistriakwn_ota_WGS84_ALL.shp')
