==> 4 add_spitoDensity_to_boundaries.ipynb <==
import libraries

import geopandas as gpd
import pandas as pd

lets count points within each Greek boundary

kap_dim = gpd.read_file('../data/oriaprokapodistriakwnota/oria_prokapodistriakwn_ota/oria_prokapodistriakwn_ota_WGS84_ALL.shp')
points = gpd.read_file('../gis/df_resolute.shp')
mapping_df = pd.read_csv('mapping_neighbourhood_LOCAL_NAME.csv')

points_mapped = pd.merge(points, mapping_df, left_on='Neighbourh', right_on='neighbourh')\
                .drop(['neighbourh'], axis='columns')
# points_mapped.head()

# points_mapped.to_csv('df_resolute_NAME_LATIN.csv', index=False)

point_count_df = pd.DataFrame(points_mapped.groupby(['NAME_LATIN']).size())\
                .reset_index()\
                .rename(columns={0:'spito_count'})

point_count_df.sort_values('spito_count', ascending=False)

#point_count_df.to_csv('df_resolute_boundary_count.csv', index=False)
# point_count_df = pd.read_csv('df_resolute_boundary_count.csv')

new_kap_dim = pd.merge(kap_dim,point_count_df,how='outer', on='NAME_LATIN')

new_kap_dim[['spito_count']] = new_kap_dim[['spito_count']].fillna(value=0)

new_kap_dim[['spito_count']] = new_kap_dim[['spito_count']].astype(int)

new_kap_dim['spito_dens']= (new_kap_dim['spito_count']/new_kap_dim['EKTASH'])*1000000

new_kap_dim.head()

# new_kap_dim.to_file(driver = 'ESRI Shapefile', filename= '../data/oriaprokapodistriakwnota/oria_prokapodistriakwn_ota/oria_prokapodistriakwn_ota_WGS84_ALL_spitoDesnity.shp')
