==> 6 data_saving.ipynb <==
import libraries

import geopandas as gpd
import pandas as pd
import matplotlib.pyplot as plt

load the kap_dim to merge them later on NAME_LATIN

kap_dim = gpd.read_file('../data/oriaprokapodistriakwnota/oria_prokapodistriakwn_ota/oria_prokapodistriakwn_ota_WGS84_ALL.shp')\
                [['NAME_LATIN','HubDist','hubdist_2', 'hubdist_3','oiki_densi', 'medianpopu','geometry']]\
                .rename(columns={'HubDist':'dist_coast','hubdist_2':'dist_beach','hubdist_3':'dist_city', 'medianpopu':'pop_densi'})

kap_dim.head()

load the cleaned spitogatos data

new_ns_frame = pd.read_csv('df_resolute_CLEANED_4IQR.csv')
new_ns_frame.sample()

len(new_ns_frame)

merge them

points_geo = pd.merge(new_ns_frame, kap_dim, on='NAME_LATIN')

points_geo.sample()

### save to csv 

# pd.DataFrame(points_geo.drop(columns='geometry')).to_csv('df_resolute_CLEANED_4IQR_4PREP.csv', index=False)

### per name find the avg price and area 

name_latin_price_area = points_geo[['PricePerM2','Area_sqm','NAME_LATIN']].groupby('NAME_LATIN').mean()\
                                                    .reset_index()

# name_latin_price_area.to_csv('name_latin_price_area.csv', index=False)

name_latin_price_area.sort_values('Area_sqm', ascending=False).head()

NAME_LATIN_meanPricePerM2 = pd.DataFrame(points_geo["PricePerM2"].groupby(points_geo['NAME_LATIN']).mean())
NAME_LATIN_meanPriceEuro = pd.DataFrame(points_geo["PriceEuro"].groupby(points_geo['NAME_LATIN']).mean())

NAME_LATIN_meanPricePerM2.sort_values('PricePerM2', ascending=False).head()

# NAME_LATIN_meanPriceEuro.sort_values('PriceEuro', ascending=False).head()
