==> 11 pricing_regression.ipynb <==
import libraries

import pandas as pd
import numpy as np
import statsmodels.api as sm
from sklearn import linear_model

data

points_geo = pd.read_csv('df_resolute_CLEANED_4IQR_4PREP_priceperm2IQR_prep.csv')
points_geo.drop(['dist_coast','dist_beach','dist_city','Year','PriceEuro','NAME_LATIN', 'longitude', 'latitude'], axis=1, inplace=True)

gps_predict_label = 'PricePerM2'

## Land 

n_points_geo = points_geo[points_geo.property_land==1]

regression

gps_hot = n_points_geo.copy()

gps_test_list = ['pop_densi', 'Area_sqm']

X = gps_hot[gps_test_list]
y = gps_hot[gps_predict_label]

scikit

lm = linear_model.LinearRegression()
model = lm.fit(X,y)

print('Intercept: \n', lm.intercept_)
print('Coefficients: \n', lm.coef_)

# R square
lm.score(X,y)

statsmodel

# Ordinary Least Squares
# add a constant
X = sm.add_constant(X)
model = sm.OLS(y, X).fit()
predictions = model.predict(X)
model.summary()

## Residential 

n_points_geo = points_geo[points_geo.property_residential==1]

regression

gps_hot = n_points_geo.copy()

gps_predict_label = 'PricePerM2'
gps_test_list = [x for x in gps_hot.columns if x!=gps_predict_label]
# gps_test_list = ['Area_sqm', 'Bedrooms', 'Bathrooms', 'age',
#                  'pop_densi', 'geohash_lat', 'geohash_lng',
#                  'near_coast','near_beach']

X = gps_hot[gps_test_list]
y = gps_hot[gps_predict_label]

scikit

lm = linear_model.LinearRegression()
model = lm.fit(X,y)

print('Intercept: \n', lm.intercept_)
print('Coefficients: \n', lm.coef_)

# predictions = lm.predict(X)
# print(predictions)

# R square
lm.score(X,y)

# # prediction with sklearn
# new_Area_sqm = 150
# new_Bedrooms = 3
# new_Bathrooms = 2
# new_age = 5
# new_pop_densi = 100
# new_property_land = 0
# new_property_residential = 1
# lista = [new_Area_sqm, new_Bedrooms,new_Bathrooms, new_age, new_pop_densi,
#          new_property_land, new_property_residential]

# pred_priceperm2 = lm.predict([lista])
# print ('Predicted Price Per M2: \n', pred_priceperm2)
# print ('Predicted Absolute Price: \n', pred_priceperm2*new_Area_sqm)

statsmodel

# # Ordinary Least Squares
# model = sm.OLS(y, X).fit()
# predictions = model.predict(X)
# model.summary()

# Ordinary Least Squares
# add a constant
X = sm.add_constant(X)
model = sm.OLS(y, X).fit()
predictions = model.predict(X)
model.summary()

## Business 

n_points_geo = points_geo[points_geo.property_business==1]

regression

gps_hot = n_points_geo.copy()

gps_predict_label = 'PricePerM2'
gps_test_list = [x for x in gps_hot.columns if x!=gps_predict_label]
# gps_test_list = ['Area_sqm', 'age','geohash_lat', 'geohash_lng',
#                  'near_coast','near_beach']

X = gps_hot[gps_test_list]
y = gps_hot[gps_predict_label]

scikit

lm = linear_model.LinearRegression()
model = lm.fit(X,y)

print('Intercept: \n', lm.intercept_)
print('Coefficients: \n', lm.coef_)

# R square
lm.score(X,y)

# Ordinary Least Squares
# add a constant
X = sm.add_constant(X)
model = sm.OLS(y, X).fit()
predictions = model.predict(X)
model.summary()
