==> 12 opportunity_finder.ipynb <==
import libraries

import pandas as pd

import project functions

from src.generalFunctions import *

data

points_mapped = pd.read_csv('df_resolute_CLEANED_4IQR_4PREP_priceperm2IQR_prep.csv')

check for prices less than 1.5 IQR just in case :P

less_iqr_filter(points_mapped, 'PricePerM2')

points_mapped.columns

# find all mean price values for each property
mean_price_land = points_mapped[points_mapped.property_land==1]['PricePerM2'].mean()
mean_price_resi = points_mapped[points_mapped.property_residential==1]['PricePerM2'].mean()
mean_price_busi = points_mapped[points_mapped.property_business==1]['PricePerM2'].mean()

land1 =  points_mapped[points_mapped.property_land==1]
land2 = more_iqr_filter(land1, 'pop_densi',0.5)
land3 = more_iqr_filter(land2, 'Area_sqm')
land4 = less_iqr_filter(land3, 'dist_city',0)
land4 = less_iqr_filter(land4, 'dist_beach',0)
land5 = land4[land4.PricePerM2<mean_price_land]
land6 = land5[land5.Year.isin([2018,2019])] # needs to be this year or last year
len(land6)

resi1 =  points_mapped[points_mapped.property_residential==1]
resi2 = less_iqr_filter(resi1, 'age', 0.2)
resi3 = less_iqr_filter(resi2, 'geohash_lat', 0.5)
resi4 = less_iqr_filter(resi3, 'dist_beach',0)
resi5 = more_iqr_filter(resi4, 'Area_sqm',0.5)
resi6 = resi5[resi5.PricePerM2<mean_price_resi]
resi7 = resi6[resi6.Year.isin([2018,2019])]  # needs to be this year or last year
len(resi7)

busi1 =  points_mapped[points_mapped.property_business==1]
busi2 = less_iqr_filter(busi1, 'age')
busi3 = less_iqr_filter(busi2, 'geohash_lat',0.5)
busi4 = busi3[busi3.PricePerM2<mean_price_busi]
busi5 = busi4[busi4.Year.isin([2018,2019])]  # needs to be this year or last year
len(busi5)

opport_df = pd.concat([land4, resi6, busi4], axis=0)

# opport_df.to_csv('df_resolute_opportunities2.csv', index=False)

# opport_df.corr()
